refine: split quote pages file; move splitCitation to CitationMarkdownFormatter
Page+Quotes.swift split into Page+Quote.swift (single + random) and
Page+QuotesList.swift. Citation splitting moved out of the view model's
private string hack into CitationMarkdownFormatter.splitCitation(from:),
which reuses the existing AST-based CiteSpanCollector.
